[style] RacingCar 파일 정리
